SLIM Plugin: Transient cursors not applied properly.  Part of Ticket #678.  Fixed chart problems #679 and flakey cursors #680.

+++ src/main/java/loci/slim/ui/IDecayGraph.java
@@ -80,5 +80,5 @@ public interface IDecayGraph {
      * @param dataStart
      * @param transientStop
      */
-    public void setStartStop(int transientStart, int dataStart, int transientStop);
+    public void setStartStop(double transientStart, double dataStart, double transientStop);
 }
